+++ functions/src/consts/env-config.const.ts
@@ -7,6 +7,12 @@ export const ENV_CONFIG = functions.config()[process.env.NODE_ENV === 'productio
   email?: {
     name: string;
     email: string;
+  },
+
+  /**
+   * Secret for email token HMAC
+   */
+  esecret: string;
   };
   gh?: {
     token: string;